Небольшие изменения в dto
