+++ src/main/java/ru/dnlkk/ratingusbackend/api/AuthApi.java
@@ -6,10 +6,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import ru.dnlkk.ratingusbackend.api.model.JWTResponseDto;
-import ru.dnlkk.ratingusbackend.api.model.UserDto;
-import ru.dnlkk.ratingusbackend.api.model.UserLoginDto;
-import ru.dnlkk.ratingusbackend.api.model.UserRegistrationDto;
+import ru.dnlkk.ratingusbackend.api.dtos.JWTRegistrationDto;
+import ru.dnlkk.ratingusbackend.api.dtos.JWTResponseDto;
+import ru.dnlkk.ratingusbackend.api.dtos.UserLoginDto;
+import ru.dnlkk.ratingusbackend.api.dtos.UserRegistrationDto;
 
 @Tag(name = "Контроллер авторизации", description = "Авторизация, регистрация")
 @RequestMapping("/auth")
@@ -27,7 +27,7 @@ public interface AuthApi {
             description = "Возвращает JWT-токен для переданных логина и пароля"
     )
     @GetMapping("/login")
-    ResponseEntity<JWTResponseDto> login(@RequestBody UserLoginDto userLoginDto);
+    ResponseEntity<JWTResponseDto> login(@RequestBody JWTResponseDto jwtResponseDto);
 
 
     @Operation(
@@ -35,5 +35,5 @@ public interface AuthApi {
             description = "Создаёт пользователя и возвращает JWT-токен"
     )
     @GetMapping("/register")
-    ResponseEntity<JWTResponseDto> register(@RequestBody UserRegistrationDto userRegistrationDto);
+    ResponseEntity<JWTResponseDto> register(@RequestBody JWTRegistrationDto jwtRegistrationDto);
 }
