+++ src/main/java/ru/dnlkk/ratingusbackend/api/ScheduleApi.java
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import ru.dnlkk.ratingusbackend.api.model.LessonScheduleDto;
+import ru.dnlkk.ratingusbackend.api.dtos.LessonScheduleDto;
 
 import java.util.List;
