+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/AnnouncementDto.java
@@ -1,0 +1,5 @@
+package ru.dnlkk.ratingusbackend.api.dtos;
+
+
+public class AnnouncementDto {
+}
