+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/DateDto.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.api.model;
+package ru.dnlkk.ratingusbackend.api.dtos;
 
 public class DateDto {
     //todo: сюда сделать timestamp
