+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/GradeDto.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.api.model;
+package ru.dnlkk.ratingusbackend.api.dtos;
 
 public class GradeDto {
     //todo: ЧТБЫ ПОСТАВИТЬ ОЦЕНКУ, нужны айди класса и айди ученика и айди предмета (вдруг кто-то одновременно в двух классах будет потом)
