+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/JWTResponseDto.java
@@ -1,0 +1,10 @@
+package ru.dnlkk.ratingusbackend.api.dtos;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class JWTResponseDto {
+    private String token;
+}
