+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/LessonDto.java
@@ -1,0 +1,4 @@
+package ru.dnlkk.ratingusbackend.api.dtos;
+
+public class LessonDto {
+}
