+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/MagazineDto.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.api.model;
+package ru.dnlkk.ratingusbackend.api.dtos;
 
 public class MagazineDto {
     //todo: нужен для представления журнала (здесь будут все ученики + их оценки и посещаемость)
