+++ src/main/java/ru/dnlkk/ratingusbackend/api/dtos/NoteDto.java
@@ -1,4 +1,4 @@
-package ru.dnlkk.ratingusbackend.api.model;
+package ru.dnlkk.ratingusbackend.api.dtos;
 
 public class NoteDto {
     //todo: айди урока, айди ученика, строка - разберётесь
