--- src/main/java/ru/dnlkk/ratingusbackend/api/model/AnnouncementDto.java
@@ -1,5 +1,0 @@
-package ru.dnlkk.ratingusbackend.api.model;
-
-
-public class AnnouncementDto {
-}
